refactor comparator to a different file for testability
